added missing i18n files / script to rebuild lang files
The script will look for all languages files in include/lang
and then copy each source file to [dst].lang.rebuilt. The
new file contains the same whitespace and comments, of course.
Then the process will scan through all known translations for
every item in the .rebuild file. If one is found, it's copied.
Else, a string in the form of "translate:..." is genereated.
The original string of the source language will be appended
as an inline comment to aid the translation process.