implemented basic versioning for addOn / core updates
The currently known addOn versions will be stored as
part of the project configuration. On every request
(FE/BE), the last known version will be compared with
the current one. If they differ, a special
'update.inc.php' will be included (if available).

The update.inc.php should die() if no update is possible.