allow controllers to handle exception on their own
It would also be possible to register the controller as the error handler, but
this would also require it to handle PHP errors and would require to un-register
the controller when the action is finished (if errors happen later on). This is
just too much work. Handling errors at one single place in your controller is
just awesome (especially if most of your actions are ajax calls and you are
required to send JSON on errors).